feat: replace regular fetch for movie details
