+++ src/shared/hooks/useGetMovieList.ts
@@ -1,9 +1,6 @@
-import { useEffect } from 'react';
-
-import useAppDispatch from './useAppDispatch.ts';
 import useAppSelector from './useAppSelector.ts';
+import useDispatchIsFetching from './useDispatchIsFetching.ts';
 import useUrl from './useUrl.ts';
-import { dataFetched } from '../../app/model/slice.ts';
 import { useGetMovieListQuery } from '../../entities/movie/api/movieApi.ts';
 import selectSearchQuery from '../../features/Search/lib/selectors/selectSearchQuery.ts';
 import { urlParams } from '../types/enums.ts';
@@ -15,7 +12,6 @@ import { urlParams } from '../types/enums.ts';
  */
 function useGetMovieList() {
   const { readUrl } = useUrl();
-  const dispatch = useAppDispatch();
 
   const query = useAppSelector(selectSearchQuery);
   const page = readUrl(urlParams.PAGE);
@@ -25,9 +21,7 @@ function useGetMovieList() {
     query,
   });
 
-  useEffect(() => {
-    dispatch(dataFetched(isFetching));
-  }, [dispatch, isFetching]);
+  useDispatchIsFetching(isFetching);
 
   return movieList;
 }
